If selectedUrl AND photo exists, do something
Here we're using `Dict.get` in a nested `case` expression.

First, we must check if our `selectedPhotoUrl` actually exists (or is it `Nothing`?) — if it does, we can then check it as a `comparable` within our Dictionary of `model.photos`.

If _that_ exists, we can pass it to `viewSelectedPhoto`. If neither `selectedPhotoUrl` or our `Dict` item exist, both return `Nothing` and we use an empty `String`.