Add toggle functionality (unfinished)
Once again this is going over my head. I'm not sure how much Elm I'll actually be able to write unaided. The toggle functions aren't making a great deal of sense to me.

And it's a LOT of code to do a very fucking simple thing. HTML can toggle it waaay easier.