Be careful with types!!!
Here we're replacing the function `viewDetailedPhoto` with a `viewFeed` function. My first attempts didn't work as each case statement **must be the same type** .. so you _could_ have `viewFeed` generate a `Photo` type for the `viewDetailedPhoto` ... but you'd have to also have an empty? `Photo` type if `Nothing` .. which is a bit of a weird way to do things.

Instead, we're simply passing an empty `text` string as a `Html Msg` instead. Simpler