Now set our data model to `Nothing`
It seems to be a tactic to make sure we're using the data model that we'd expect, _before_ loading in that data from the server.

This allows us to build out our program without worrying about validating and extracting `json` (which we've already tested locally with our `decoder`).

So, build your decoder, use the model you'd expect to see on successful decoding, then load it in and test things out.