Add some harsh notes (but I don't mind!)
I'm good at a basic prototype level, but then I'm happy to handover to people who are much more experienced and suited to the task than me. I've tried! A lot!

I could go back to HTDP (which I might do in the future) but I'm more comfortable at the high-level stage rather than in the weeds with algorithms and such.

The very real fact is, like piano, if you only tinker, you only get simple pop songs. You've really got to be incredibly serious and perseverant (and hopefully have the right apptitude) for CompSci. Like someone once said to me:

"If you're not working 12 hours a day, you're gonna struggle"

The same goes with piano as goes with code.