Consolidate the notes, there's still too much!
Writing can help to rejig the memory and consolidate some thoughts, but it all takes time — deep learning takes a lot of effort, time, and energy. And of course, it's all stuff that on the micro I understand but in the macro I find hard to piece together. :alarm_clock: :poop:

As someone once said to me, I'm never going to set the world on fire with programming. So focus in more of your time on stuff that will! :fire: