Made some changes to the model
If we just have our `type Custom` as `List Entry`, it doesn't register as a `Entries` type (why, I don't know).

You seem to have to need to add the "parent" type `Entries` in front of the original `List Entry` to make it work. Also wrap the list in brackets.