Kill a bug (comparables)
I've changed to a `case` as **you can't compare a type with children types** — you can't do `NoEntries == [Entry 123 "string"]` as they're not comparables. You _can_ compare `NoEntries == NoEntries` (and maybe there's a way to do it with `not` and a second branch.

`case` seems a bit easier anyway.

## The original `if` statement problem, for future notice ...

When running the `if` expression I was seeing the following error:

```terminal
I was expecting to see more of this `if` expression, but I got stuck here:

89|         if isEntry entries then
90|             ul [ class "entry-list" ]
91|                 viewEntries entries
92|             , viewForm currentEntry
                ^
I was expecting to see the else keyword next.
```

Seems you'd need to wrap the children with a `div [] [ ... ]` _within_ the `if` branch.