Add a page for future reference
We're no longer using the `Entries` type as it's not a good example of custom types in this instance, as it doesn't buy us anything over a simple `List Entry` — it's either a `[]` (empty) or `[Entry 1 "something"]`  (single) or multiple entries.

If we want to add safety, i.e. we don't know if the server will give us a `List Entry` (or even a list) we can use `Maybe`.

Making impossible states impossible is where custom types really shine. That's when you have interconnected data that could go wrong otherwise.

Mostly my data is dumb, but if for instance you had `tag` which could have many `parents`, that might be a good area for custom types.