Ugh ugh and ugh
Typed is helpful in some ways, but all these fucking `Maybe`s are getting on my nerves. You _really_ have to think out how you're structuring your calls, and I can see now why the `Result` in the Elm Guide spits out the `String.toFloat` _within_ the very first `case __ of` statement.

Otherwise you keep on having to check and extract the damn thing in multiple places.

There must be an easier way to check x amount of decimal points in a Float. I think I need to be EXTRA careful from the start how these "black boxes" are processed, and the original inputs and methods of collecting the data.

It's a very annoyingly slow process.