Tidying up the base custom type
We've got an ID, a song time, song name. It's not essential to have a custom type `Album`, as a `List Song` would probably do. It's not a collection, but we'll need a form and a view. A `Tuple` might not be the best format for the time, but it's all in one record field.