How can I reduce the surface area of errors?
Basically, anything that is complicated or involves a lot of state changing of data, is going to be tricky to do if you're not coding very often. If the goal is to prototype rapidly, focus 50% time on marketing and sales, why make life harder than it needs to be?

Simplify. Simplify. Simplify.

Sketch out your problem and gauge how difficult it actually is.

Then simplify some more!